Refactored Animations into the AnimatorUtil, and Refactored the Dice utility functions into a DiceUtil class. Main Activity is much cleaner and easier to read now.
