+++ app/src/main/java/com/personal/peter/easydice/MainActivity.java
@@ -1,7 +1,5 @@
 package com.personal.peter.easydice;
 
-import android.animation.Animator;
-import android.animation.AnimatorInflater;
 import android.content.Context;
 import android.content.SharedPreferences;
 import android.support.v7.app.AppCompatActivity;
@@ -12,10 +10,8 @@ import android.view.View;
 import android.widget.Button;
 import android.widget.EditText;
 import android.widget.TextView;
-import android.widget.Toast;
 import java.util.ArrayList;
 import java.util.HashMap;
-import java.util.List;
 
 
 import butterknife.BindView;
@@ -26,16 +22,8 @@ public class MainActivity extends AppCompatActivity {
     private static final String PREFS_FILE = "com.personal.peter.easydice.preferences";
     private SharedPreferences mSharedPreferences;
     private SharedPreferences.Editor mEditor;
-
-    private static final String KEY_D4 = "KEY_D4";
-    private static final String KEY_D6 = "KEY_D6";
-    private static final String KEY_D8 = "KEY_D8";
-    private static final String KEY_D10 = "KEY_D10";
-    private static final String KEY_D12 = "KEY_D12";
-    private static final String KEY_D20 = "KEY_D20";
     private static final String KEY_TOTAL = "KEY_TOTAL";
     private static final String KEY_RESULTS = "KEY_RESULTS";
-
     private ArrayList<String> mKeys;
     private HashMap<String, String> mKeyMap;
     private HashMap<String, EditText> mTextMap;
@@ -76,8 +64,6 @@ public class MainActivity extends AppCompatActivity {
     @BindView(R.id.rollArrayLabel)
     TextView mRollArrayLabel;
 
-   private int dieNum;
-
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
@@ -85,6 +71,7 @@ public class MainActivity extends AppCompatActivity {
         setContentView(R.layout.activity_main);
         ButterKnife.bind(this);
         animatorUtil = new AnimatorUtil(this);
+        final DiceUtil diceUtil = new DiceUtil(this, mTotalTextView, mResultTextView);
 
         mKeys = new ArrayList<>(6);
         mKeys.add("D4");
@@ -128,16 +115,15 @@ public class MainActivity extends AppCompatActivity {
             @Override
             public void onClick(View v) {
                 animatorUtil.animateButton(mD4Button);
-                takeInputAndRoll(mD4EditText.getText().toString(), 4);
+                diceUtil.takeInputAndRoll(mD4EditText.getText().toString(), 4);
             }
         });
-
 
         mD6Button.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View v) {
                 animatorUtil.animateButton(mD6Button);
-                takeInputAndRoll(mD6EditText.getText().toString(), 6);
+                diceUtil.takeInputAndRoll(mD6EditText.getText().toString(), 6);
             }
         });
 
@@ -145,7 +131,7 @@ public class MainActivity extends AppCompatActivity {
             @Override
             public void onClick(View v) {
                 animatorUtil.animateButton(mD8Button);
-                takeInputAndRoll(mD8EditText.getText().toString(), 8);
+                diceUtil.takeInputAndRoll(mD8EditText.getText().toString(), 8);
             }
         });
 
@@ -153,7 +139,7 @@ public class MainActivity extends AppCompatActivity {
             @Override
             public void onClick(View v) {
                 animatorUtil.animateButton(mD10Button);
-                takeInputAndRoll(mD10EditText.getText().toString(), 10);
+                diceUtil.takeInputAndRoll(mD10EditText.getText().toString(), 10);
             }
         });
 
@@ -161,7 +147,7 @@ public class MainActivity extends AppCompatActivity {
             @Override
             public void onClick(View v) {
                 animatorUtil.animateButton(mD12Button);
-                takeInputAndRoll(mD12EditText.getText().toString(), 12);
+                diceUtil.takeInputAndRoll(mD12EditText.getText().toString(), 12);
             }
         });
 
@@ -169,10 +155,9 @@ public class MainActivity extends AppCompatActivity {
             @Override
             public void onClick(View v) {
                 animatorUtil.animateButton(mD20Button);
-                takeInputAndRoll(mD20EditText.getText().toString(), 20);
+                diceUtil.takeInputAndRoll(mD20EditText.getText().toString(), 20);
             }
         });
-
     }
 
     @Override
@@ -185,47 +170,6 @@ public class MainActivity extends AppCompatActivity {
         mEditor.putString(KEY_TOTAL, mTotalTextView.getText().toString());
         mEditor.putString(KEY_RESULTS, mResultTextView.getText().toString());
         mEditor.apply();
-    }
-
-
-    private void rollAndSet(Dice die) {
-        die.rollDice();
-        mTotalTextView.setText(die.getTotal() + "");
-        mResultTextView.setText(getResultsString(die.getResults()));
-        animatorUtil.animateText(mTotalTextView);
-        animatorUtil.animateText(mResultTextView);
-    }
-
-    private String getResultsString(List<Integer> results) {
-        List<String> stringList = new ArrayList<>();
-        for (int value : results) stringList.add(Integer.toString(value));
-        StringBuilder result = new StringBuilder();
-        for (String string : stringList) {
-            result.append(string);
-            result.append(", ");
-        }
-        return result.length() > 0 ? result.substring(0, result.length() - 2) : "";
-    }
-
-
-    private boolean isValidInput(int input) {
-        return (input > 0 && input < 61);
-    }
-
-    private void displayToast() {
-        Toast.makeText(this, "Must be between 1 and 60 dice.", Toast.LENGTH_SHORT).show();
-    }
-
-    private void takeInputAndRoll(String string, int dieSize) {
-        try {
-            dieNum = Integer.parseInt(string);
-            if (isValidInput(dieNum)) {
-                Dice die = new Dice(dieNum, dieSize, 0, 0);
-                rollAndSet(die);
-            } else displayToast();
-        } catch (NumberFormatException e) {
-            displayToast();
-        }
     }
 
     @Override
@@ -247,7 +191,6 @@ public class MainActivity extends AppCompatActivity {
             text.setText("1");
             animatorUtil.animateText(text);
         }
-
         return true;
     }
 }
